refactor: factor out components, scope syncComplete, remove debug isMatch, pluralize
