+++ client/src/feat/my-passport-form-sync/hooks/useIsMatch.ts
@@ -7,17 +7,10 @@ export const useIsMatch = () => {
 	const userContext = useContext(UserContext);
 	const searchParams = new URLSearchParams(window.location.search);
 
-	const isMatch = () => {
-		if (import.meta.env.DEV) {
-			return searchParams.has("debug", "myPassportFormSync");
-		}
-
-		return (
-			Boolean(authnContext().keycloak?.token) &&
-			!userContext().syncComplete &&
-			!searchParams.has("view", "sync")
-		);
-	};
+	const isMatch = () =>
+		Boolean(authnContext().keycloak?.token) &&
+		!userContext().syncComplete &&
+		!searchParams.has("view", "sync");
 
 	return isMatch;
 };
