+++ client/src/feat/my-passport-form-sync/hooks/useToaster.tsx
@@ -1,20 +1,12 @@
 import { toaster } from "@kobalte/core";
 import { useQueryClient } from "@tanstack/solid-query";
 import { AuthnContext } from "core/context/authn";
-import { useI18n } from "core/context/i18n";
 import { UserContext } from "core/context/user";
 import { useContext } from "core/context/utils";
 import { invariant } from "es-toolkit";
-import type { resources } from "feat/my-passport-form-sync/resources/i18n/en-US";
 import { queryKeys } from "feat/my-passport-form-sync/resources/query-keys";
-import {
-	createReaction,
-	createSignal,
-	type Component,
-	type ParentProps,
-} from "solid-js";
-import { Toast, ToastContent, ToastDescription, ToastTitle } from "ui/toast";
-import { withI18n } from "../resources";
+import { SyncToast } from "feat/my-passport-form-sync/ui/sync-toast";
+import { createReaction, createSignal } from "solid-js";
 
 export const useToaster = () => {
 	const userContext = useContext(UserContext);
@@ -40,12 +32,10 @@ export const useToaster = () => {
 		invariant(numberOfApplications, "Number of applications is falsy");
 
 		const id = toaster.show(props => (
-			<>
-				<MyPassportFormSyncToast
-					toastId={props.toastId}
-					numberOfApplications={numberOfApplications}
-				/>
-			</>
+			<SyncToast
+				toastId={props.toastId}
+				numberOfApplications={numberOfApplications}
+			/>
 		));
 
 		setToastId(id);
@@ -55,26 +45,3 @@ export const useToaster = () => {
 
 	return toastId;
 };
-
-interface MyPassportFormSyncToastProps {
-	toastId: number;
-	numberOfApplications: number;
-}
-
-const MyPassportFormSyncToast: Component<
-	ParentProps<MyPassportFormSyncToastProps>
-> = withI18n(props => {
-	const t = useI18n<typeof resources>();
-
-	return (
-		<Toast toastId={props.toastId}>
-			<ToastContent>
-				<ToastTitle>{t("toastTitle")}</ToastTitle>
-
-				<ToastDescription>
-					{t("toastDescription", props.numberOfApplications)}
-				</ToastDescription>
-			</ToastContent>
-		</Toast>
-	);
-});
