+++ client/src/feat/my-passport-form-sync/my-passport-form-sync.tsx
@@ -1,16 +1,6 @@
-import { MyPassportForm } from "core/constants/my-passport-form-route.enum";
 import { useI18n } from "core/context/i18n";
-import { generatePath } from "core/utils";
-import { SquareArrowOutUpRight } from "lucide-solid";
-import {
-	createSignal,
-	For,
-	Show,
-	type Component,
-	type ParentProps,
-} from "solid-js";
+import { For, Show } from "solid-js";
 import { Button } from "ui/button";
-import { Checkbox, CheckboxControl } from "ui/checkbox";
 import {
 	Dialog,
 	DialogContent,
@@ -19,10 +9,9 @@ import {
 	DialogHeader,
 	DialogTitle,
 } from "ui/dialog";
-import { Typography } from "ui/typography";
-import { UUID } from "uuidv7";
 import { useMyPassportFormSync } from "./hooks";
 import type { resources } from "./resources/i18n/en-US";
+import { PassportApplicationLi } from "./ui/passport-application-li";
 
 export const MyPassportFormSync = () => {
 	const {
@@ -52,7 +41,7 @@ export const MyPassportFormSync = () => {
 							<For each={data.publicApplications()}>
 								{application => (
 									<>
-										<PassportApplicationListItem
+										<PassportApplicationLi
 											uuid={application.uuid}
 											automergeUrl={
 												application.automergeUrl
@@ -77,82 +66,4 @@ export const MyPassportFormSync = () => {
 			</Dialog>
 		</Show>
 	);
-};
-
-export interface PassportApplicationListItemProps {
-	uuid: string;
-	automergeUrl: string;
-}
-
-const PassportApplicationListItem: Component<
-	ParentProps<PassportApplicationListItemProps>
-> = props => {
-	const t = useI18n<typeof resources>();
-
-	const [isChecked, setIsChecked] = createSignal(false);
-
-	const getHref = () => {
-		const url = new URL(location.origin);
-		url.hash = location.hash;
-
-		const searchParams = new URLSearchParams({
-			automergeUrl: props.automergeUrl,
-			view: "sync",
-		});
-
-		url.pathname = generatePath(MyPassportForm.Edit, {
-			uuid: props.uuid,
-		});
-		url.search = searchParams.toString();
-
-		return url.href;
-	};
-
-	return (
-		<li class="flex items-center justify-between p-4">
-			<div class="flex space-x-2 items-center">
-				<Typography>
-					{t("listItemDescription", toDate(props.uuid))}
-				</Typography>
-			</div>
-
-			<div class="flex space-x-4 items-center">
-				<Checkbox
-					name={props.uuid}
-					value={props.automergeUrl}
-					checked={isChecked()}
-					onChange={setIsChecked}>
-					<CheckboxControl />
-				</Checkbox>
-
-				<a
-					href={getHref()}
-					target="_blank"
-					rel="noopener noreferrer"
-					class="focus:outline-none focus-visible:outline-none"
-					tabIndex={-1}>
-					<Button tabIndex={0} variant="secondary" size="icon">
-						<SquareArrowOutUpRight />
-					</Button>
-				</a>
-			</div>
-		</li>
-	);
-};
-
-// from: https://gist.github.com/wllmsash/bcb337ce0662ed044012e2d7170f7ed0
-const toDate = (uuid: string) => {
-	const timestampBytes = new Uint8Array(8);
-	timestampBytes.set(
-		// first 6 bytes are timestamp
-		new Uint8Array(UUID.parse(uuid).bytes.buffer.slice(0, 6)),
-		// leave first 2 bytes empty
-		// `getBigUint64` reads 8 bytes
-		2,
-	);
-
-	// unix timestamp
-	const timestampMs = new DataView(timestampBytes.buffer).getBigUint64(0);
-
-	return new Date(Number(timestampMs));
 };
